requirements.txt
asgiref==3.8.1
<<<<<<< HEAD
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
channels
daphne
=======
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
defusedxml==0.7.1
Django==5.2.3
django-allauth==0.57.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
google-auth==2.40.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python3-openid==3.2.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
zope.interface==7.2
psycopg2-binary
>>>>>>> eb0a0dc1152f52687a39c643193460d5aa501fb3
